Minercantile items are now craftable
Barter shop and bancomatic are craftable by anyone but to administrate the barter shop the player must have the shop privilege. (and ask to get it)
Anyway a non priviligied user can use his barter shop, but it will be limited to sell/bye objects from the General group that is administrated by players with the shop privilege.